+++ argilla-frontend/components/features/annotation/container/fields/RecordMetadataInfo.vue
@@ -92,13 +92,13 @@ export default {
 <style lang="scss" scoped>
 .table-info {
   background: palette(white);
-  min-width: 460px;
+  min-width: 440px;
   border-radius: $border-radius-s;
   overflow: hidden;
   @include font-size(13px);
   box-shadow: $shadow;
   &__content {
-    max-height: 320px;
+    max-height: 280px;
     overflow: auto;
   }
   &__header {
@@ -112,6 +112,7 @@ export default {
     .--table & {
       background: $black-6;
       border-top: 1px solid $black-10;
+      border-bottom: 1px solid $black-10;
     }
   }
   &__row {
@@ -120,7 +121,7 @@ export default {
     gap: $base-space * 2;
     padding: $base-space $base-space * 2;
     .--table & {
-      border-top: 1px solid $black-10;
+      border-bottom: 1px solid $black-10;
       &:nth-child(even) {
         background: $black-4;
       }
@@ -135,6 +136,9 @@ export default {
     flex-basis: 44%;
     margin: 0;
     @include truncate;
+    .--intro & {
+      flex-basis: 100px;
+    }
   }
   &__value {
     display: flex;
@@ -145,6 +149,9 @@ export default {
     margin: 0;
     font-family: $quaternary-font-family;
     word-break: break-all;
+    .--intro & {
+      flex-basis: 100%;
+    }
   }
   &__copy {
     opacity: 0;