+++ argilla-frontend/jest.setup.ts
@@ -59,6 +59,31 @@ config.mocks = {
   },
 };
 
+export class IntersectionObserverMock {
+  root = null;
+  rootMargin = "";
+  thresholds = [];
+
+  disconnect() {
+    return null;
+  }
+
+  observe() {
+    return null;
+  }
+
+  takeRecords() {
+    return [];
+  }
+
+  unobserve() {
+    return null;
+  }
+}
+
+window.IntersectionObserver = IntersectionObserverMock;
+global.IntersectionObserver = IntersectionObserverMock;
+
 export default async () => {
   const nuxt = await buildNuxt();