+++ argilla-sdk/src/argilla_sdk/_helpers/_iterator.py
@@ -1,0 +1,37 @@
+# Copyright 2024-present, Argilla, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Generic, List, TypeVar
+
+__all__ = ["GenericIterator"]
+
+Item = TypeVar("Item")
+
+
+class GenericIterator(Generic[Item]):
+    """Generic iterator for any collection of items."""
+
+    def __init__(self, collection: List[Item]):
+        self._collection = [v for v in collection]
+        self._index = 0
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        if self._index < len(self._collection):
+            result = self._collection[self._index]
+            self._index += 1
+            return result
+        raise StopIteration