+++ argilla-server/src/argilla_server/models/mixins.py
@@ -14,19 +14,22 @@
 
 from datetime import datetime
 from typing import TYPE_CHECKING, Any, Dict, List, Set, TypeVar, Union
+from uuid import UUID
 
-from sqlalchemy import sql
+from sqlalchemy import select, sql
 from sqlalchemy.dialects.mysql import insert as mysql_insert
 from sqlalchemy.dialects.postgresql import insert as postgres_insert
 from sqlalchemy.dialects.sqlite import insert as sqlite_insert
 from sqlalchemy.engine.default import DefaultExecutionContext
+from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import Mapped, mapped_column
+from sqlalchemy.sql.base import ExecutableOption
 from typing_extensions import Self
 
+from argilla_server.errors.future import NotFoundError
 from argilla_server.pydantic_v1 import BaseModel
 
 if TYPE_CHECKING:
-    from sqlalchemy.ext.asyncio import AsyncSession
     from sqlalchemy.orm import InstrumentedAttribute
     from sqlalchemy.sql.elements import BinaryExpression
 
@@ -64,7 +67,7 @@ class CRUDMixin:
 
     @classmethod
     async def create(
-        cls, db: "AsyncSession", schema: Union[Schema, None] = None, autocommit: bool = True, **kwargs: Any
+        cls, db: AsyncSession, schema: Union[Schema, None] = None, autocommit: bool = True, **kwargs: Any
     ) -> Self:
         _values = _schema_or_kwargs(schema, kwargs)
         instance = cls()
@@ -72,19 +75,34 @@ class CRUDMixin:
         return await instance.save(db, autocommit)
 
     @classmethod
-    async def read(cls, db: "AsyncSession", id: Any, key: str = "id") -> Union[Self, None]:
-        params = {key: id}
-        return await cls.read_by(db, **params)
+    async def get(cls, db: AsyncSession, id: UUID, options: List[ExecutableOption] = []) -> Union[Self, None]:
+        return (await db.execute(select(cls).filter_by(id=id).options(*options))).scalar_one_or_none()
 
     @classmethod
-    async def read_by(cls, db: "AsyncSession", **params: Any) -> Union[Self, None]:
-        query = sql.select(cls).filter_by(**params)
-        result = await db.execute(query)
-        return result.scalars().unique().one_or_none()
+    async def get_or_raise(cls, db: AsyncSession, id: UUID, options: List[ExecutableOption] = []) -> Self:
+        instance = await cls.get(db, id, options)
+        if instance is not None:
+            return instance
+
+        raise NotFoundError(f"{cls.__name__} with id `{id}` not found")
+
+    @classmethod
+    async def get_by(cls, db: AsyncSession, **conditions) -> Union[Self, None]:
+        return (await db.execute(select(cls).filter_by(**conditions))).scalar_one_or_none()
+
+    @classmethod
+    async def get_by_or_raise(cls, db: AsyncSession, **conditions) -> Self:
+        instance = await cls.get_by(db, **conditions)
+        if instance is not None:
+            return instance
+
+        conditions_str = ", ".join([f"{key}={value}" for key, value in conditions.items()])
+
+        raise NotFoundError(f"{cls.__name__} not found filtering by {conditions_str}")
 
     async def update(
         self,
-        db: "AsyncSession",
+        db: AsyncSession,
         schema: Union[Schema, None] = None,
         replace_dict: bool = False,
         autocommit: bool = True,
@@ -97,7 +115,7 @@ class CRUDMixin:
     @classmethod
     async def update_many(
         cls,
-        db: "AsyncSession",
+        db: AsyncSession,
         objects: List[Dict[str, Any]],
         autocommit: bool = True,
     ) -> None:
@@ -112,7 +130,7 @@ class CRUDMixin:
     @classmethod
     async def upsert_many(
         cls,
-        db: "AsyncSession",
+        db: AsyncSession,
         objects: List[Union[Schema, dict]],
         constraints: List["InstrumentedAttribute[Any]"],
         autocommit: bool = True,
@@ -146,7 +164,7 @@ class CRUDMixin:
     @classmethod
     async def upsert(
         cls,
-        db: "AsyncSession",
+        db: AsyncSession,
         schema: Union[Schema, dict],
         constraints: List["InstrumentedAttribute[Any]"],
         autocommit: bool = True,
@@ -155,7 +173,7 @@ class CRUDMixin:
 
         return upserted[0]
 
-    async def delete(self, db: "AsyncSession", autocommit: bool = True) -> Self:
+    async def delete(self, db: AsyncSession, autocommit: bool = True) -> Self:
         await db.delete(self)
         if autocommit:
             await db.commit()
@@ -163,7 +181,7 @@ class CRUDMixin:
 
     @classmethod
     async def delete_many(
-        cls, db: "AsyncSession", params: List["BinaryExpression"], autocommit: bool = True
+        cls, db: AsyncSession, params: List["BinaryExpression"], autocommit: bool = True
     ) -> List[Self]:
         delete_stmt = sql.delete(cls).filter(*params).returning(cls)
         result = await db.execute(delete_stmt)
@@ -171,7 +189,7 @@ class CRUDMixin:
             await db.commit()
         return result.scalars().all()
 
-    async def save(self, db: "AsyncSession", autocommit: bool = True) -> Self:
+    async def save(self, db: AsyncSession, autocommit: bool = True) -> Self:
         db.add(self)
         if autocommit:
             await db.commit()