+++ argilla-server/src/argilla_server/validators/responses.py
@@ -14,7 +14,8 @@
 
 from typing import Union
 
-from argilla_server.enums import ResponseStatus
+from argilla_server.enums import QuestionType, ResponseStatus
+from argilla_server.errors.future import UnprocessableEntityError
 from argilla_server.models import Record
 from argilla_server.schemas.v1.responses import ResponseCreate, ResponseUpdate, ResponseUpsert
 from argilla_server.validators.response_values import ResponseValueValidator
@@ -36,19 +37,23 @@ class ResponseValidator:
 
     def _validate_values_are_present_when_submitted(self) -> None:
         if self._is_submitted_response and not self._response_change.values:
-            raise ValueError("missing response values for submitted response")
+            raise UnprocessableEntityError("missing response values for submitted response")
 
     def _validate_required_questions_have_values(self, record: Record) -> None:
         for question in record.dataset.questions:
             if self._is_submitted_response and question.required and question.name not in self._response_change.values:
-                raise ValueError(f"missing response value for required question with name={question.name!r}")
+                raise UnprocessableEntityError(
+                    f"missing response value for required question with name={question.name!r}"
+                )
 
     def _validate_values_have_configured_questions(self, record: Record) -> None:
         question_names = [question.name for question in record.dataset.questions]
 
         for value_question_name in self._response_change.values or []:
             if value_question_name not in question_names:
-                raise ValueError(f"found response value for non configured question with name={value_question_name!r}")
+                raise UnprocessableEntityError(
+                    f"found response value for non configured question with name={value_question_name!r}"
+                )
 
     def _validate_values(self, record: Record) -> None:
         if not self._response_change.values: