+++ argilla-server/tests/unit/api/v1/datasets/records/records_bulk/test_dataset_records_bulk_with_responses.py
@@ -309,10 +309,10 @@ class TestDatasetRecordsBulkWithResponses:
                                 "user_id": str(user.id),
                                 "values": {"other-question": {"value": "wrong value"}},
                                 "status": "draft",
-                            }
+                            },
                         ],
                     },
-                ]
+                ],
             },
         )