+++ argilla-server/CHANGELOG.md
@@ -16,6 +16,12 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+### Removed
+
+- Removed all API v0 endpoints. ([#4852](https://github.com/argilla-io/argilla/pull/4852))
+
+## [1.29.0](https://github.com/argilla-io/argilla/compare/v1.28.0...v1.29.0)
+
 ### Added
 
 - Added support for rating questions to include `0` as a valid value. ([#4858](https://github.com/argilla-io/argilla/pull/4858))
@@ -32,9 +38,9 @@ These are the section headers that we use:
 - Added `GET /api/v1/version` endpoint to get the current Argilla version. ([#162](https://github.com/argilla-io/argilla-server/pull/162))
 - Added `GET /api/v1/status` endpoint to get Argilla service status. ([#165](https://github.com/argilla-io/argilla-server/pull/165))
 
-### Removed
+## Fixed
 
-- Removed all API v0 endpoints. ([#4852](https://github.com/argilla-io/argilla/pull/4852))
+- Fixed error when fetching record with metadata values with visibility restrictions. ([#4906](https://github.com/argilla-io/argilla/pull/4906))
 
 ## [1.28.0](https://github.com/argilla-io/argilla-server/compare/v1.27.0...v1.28.0)