+++ argilla/CHANGELOG.md
@@ -16,9 +16,15 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+## [1.29.0](https://github.com/argilla-io/argilla/compare/v1.28.0...v1.29.0)
+
 ### Added
 
 - Added support for rating questions to include `0` as a valid value. ([#4860](https://github.com/argilla-io/argilla/pull/4860))
+- Added support for Python 3.12. ([#4837](https://github.com/argilla-io/argilla/pull/4837))
+- Added search by field in the `FeedbackDataset` UI search. ([#4746](https://github.com/argilla-io/argilla/issues/4746))
+- Added record metadata info in the `FeedbackDataset` UI. ([#4851](https://github.com/argilla-io/argilla/pull/4851))
+- Added highlight on search results in the `FeedbackDataset` UI. ([#4747](https://github.com/argilla-io/argilla/issues/4747))
 
 ### Fixed